requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
filelock==3.4.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mechanize==0.4.5
oauthlib==3.1.1
packaging==21.3
phonenumbers==8.12.17
Pillow==8.0.1
platformdirs==2.4.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.1
sqlparse==0.4.2
tzdata==2021.5
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
whitenoise==5.3.0
wincertstore==0.2
